change INT to BIGINT

+++ models.js
@@ -3,12 +3,12 @@ const { DataTypes } = require("sequelize");
 
 const User = sequelize.define("user", {
   id: {
-    type: DataTypes.INTEGER,
+    type: DataTypes.BIGINT,
     primaryKey: true,
     unique: true,
   },
   username: { type: DataTypes.STRING },
-  chatId: { type: DataTypes.INTEGER },
+  chatId: { type: DataTypes.BIGINT },
   pidorCount: { type: DataTypes.INTEGER, defaultValue: 0 },
 });
